[bot追加] リファクタ&ローカル実行用のボットひな型実装
