+++ .github/copilot-instrucions.md
@@ -8,12 +8,17 @@
     - main.* ファイルをエントリポイントとし、controller.* に渡す構成を採用
       - バッチやツールなどServiceなどで解決する場合
     - 必ずインターフェイスとクラスを利用
+      - ～サービスのクラス
     - 必ず型定義
 - フロントエンド
   - typescript を使用し、UIコンポーネント中心のシンプルなSPA構成を採用
     - index.ts がエントリポイント
     - 必ず型定義
 </ProjectLayout>
+<Code>
+- 一般的な高級言語
+  - エントリポイントの関数を除き、グローバル領域にはクラス定義とimport文のみを入れる
+</Code>
 <Function>
 - 関数名直下での引数・リターン付きのコメントを入れる
 </Function>
