+++ bots/batch/ris-bot-birthday-manjuuu/main.py
@@ -1,8 +1,8 @@
 import os
 import discord
 from dotenv import load_dotenv
-from services.birthday_notification import birthday_notification_main
-from services.youtube_pickup_song import youtube_pickup_song_main
+from services.birthday_notification_service import BirthdayNotificationService
+from services.youtube_pickup_song_service import YouTubePickupSongService
 
 # env読み込み
 load_dotenv()
@@ -16,8 +16,10 @@ client = discord.Client(intents=intents)
 
 @client.event
 async def on_ready():
-    await birthday_notification_main(client)
-    await youtube_pickup_song_main(client)
+    birthday_service = BirthdayNotificationService()
+    youtube_service = YouTubePickupSongService()
+    await birthday_service.birthday_notification_main(client)
+    await youtube_service.add_youTube_playlist_main(client)
     await client.close()
 
 if __name__ == "__main__":
